keyboard: Move source files to internal directory
